+++ XIVAuras/Helpers/Data.cs
@@ -19,14 +19,16 @@ namespace XIVAuras.Helpers
         public ushort Icon;
         public byte MaxStacks;
         public uint[] ComboId;
+        public CombatType CombatType;
 
-        public TriggerData(string name, uint id, ushort icon, byte maxStacks = 0, uint[]? comboId = null)
+        public TriggerData(string name, uint id, ushort icon, byte maxStacks = 0, uint[]? comboId = null, CombatType combatType = CombatType.PvE)
         {
             Name = name;
             Id = id;
             Icon = icon;
             MaxStacks = maxStacks;
             ComboId = comboId ?? new uint[0];
+            CombatType = combatType;
         }
     }